+++ pyuavcan/transport/can/media/socketcan/_socketcan.py
@@ -195,7 +195,7 @@ class SocketCANMedia(Media):
 
     def _read_frame(self, ts_mono_ns: int) -> typing.Tuple[Timestamp, Envelope]:
         while True:
-            data, ancdata, msg_flags, _addr = self._sock.recvmsg(
+            data, ancdata, msg_flags, _addr = self._sock.recvmsg(  # type: ignore
                 self._native_frame_size, self._ancillary_data_buffer_size
             )
             assert msg_flags & socket.MSG_TRUNC == 0, "The data buffer is not large enough"
@@ -295,7 +295,7 @@ _CAN_ERR_FLAG = 0x20000000
 _CAN_EFF_MASK = 0x1FFFFFFF
 
 
-def _make_socket(iface_name: str, can_fd: bool) -> socket.SocketType:
+def _make_socket(iface_name: str, can_fd: bool) -> socket.socket:
     s = socket.socket(socket.PF_CAN, socket.SOCK_RAW, socket.CAN_RAW)  # type: ignore
     try:
         s.bind((iface_name,))